LayoutTests/editing/shadow/selection-of-orphan-shadowroot.html

<!DOCTYPE html>
<html>
<head>
<script src="../../fast/js/resources/js-test-pre.js"></script>
<script src="../../fast/dom/resources/event-sender-util.js"></script>
<script src="../../fast/dom/resources/shadow-test-driver.js"></script>
<script src="../../fast/dom/shadow/resources/polyfill.js"></script>
</head>
<body>

<p>Nodes of the selection for an orphan shadow root should return null.</p>

<pre id="console"></pre>

<script>
var host = document.createElement('div');
var shadowRoot = new WebKitShadowRoot(host);
shadowRoot.innerHTML = "Kotori Otonashi"

var selection = shadowRoot.selection;

shouldBe('selection.anchorNode', 'null');
shouldBe('selection.focusNode', 'null');
shouldBe('selection.baseNode', 'null');
shouldBe('selection.extentNode', 'null');

var successfullyParsed = true;
</script>
<script src="../../fast/js/resources/js-test-post.js"></script>
</body>

</html>
